Add core namespace and improve events.
This adds a core namespace for dispatching high level endpoint commands to
other namespaces. It then applies the args given to another dispatch function
in each namspace.

Adds an info and health function to core which hit the corresponding endpoints
in the Sensu API.

Also semi-settled on a format for passing the host around. Every function
accepts a host parameter, but it's the first in core/dispatch, meaning you can
easily partially apply it and then not have to worry about it again.